Moved Request State functions to utils
