+++ e2e-polybft/e2e/cardano_test.go
@@ -3,13 +3,11 @@ package e2e
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"math/big"
 	"net/http"
 	"path"
-	"strings"
 	"sync"
 	"testing"
 	"time"
@@ -260,7 +258,7 @@ for_loop:
 			continue
 		}
 
-		var responseModel BridgingRequestStateResponse
+		var responseModel cardanofw.BridgingRequestStateResponse
 
 		err = json.Unmarshal(resBody, &responseModel)
 		if err != nil {
@@ -360,7 +358,7 @@ func TestE2E_InvalidScenarios(t *testing.T) {
 		requestURL := fmt.Sprintf(
 			"%s/api/BridgingRequestState/Get?chainId=%s&txHash=%s", apiURL, "prime", txHash)
 
-		state, err := WaitForRequestState(InvalidState, ctx, "prime", txHash, requestURL, apiKey, 300)
+		state, err := cardanofw.WaitForRequestState(InvalidState, ctx, requestURL, apiKey, 300)
 		require.NoError(t, err)
 		require.Equal(t, state, InvalidState)
 	})
@@ -388,7 +386,7 @@ func TestE2E_InvalidScenarios(t *testing.T) {
 			requestURL := fmt.Sprintf(
 				"%s/api/BridgingRequestState/Get?chainId=%s&txHash=%s", apiURL, "prime", txHash)
 
-			state, err := WaitForRequestState(InvalidState, ctx, "prime", txHash, requestURL, apiKey, 300)
+			state, err := cardanofw.WaitForRequestState(InvalidState, ctx, requestURL, apiKey, 300)
 			require.NoError(t, err)
 			require.Equal(t, state, InvalidState)
 		}
@@ -447,7 +445,7 @@ func TestE2E_InvalidScenarios(t *testing.T) {
 			requestURL := fmt.Sprintf(
 				"%s/api/BridgingRequestState/Get?chainId=%s&txHash=%s", apiURL, "prime", txHashes[i])
 
-			state, err := WaitForRequestState(InvalidState, ctx, "prime", txHashes[i], requestURL, apiKey, 300)
+			state, err := cardanofw.WaitForRequestState(InvalidState, ctx, requestURL, apiKey, 300)
 			require.NoError(t, err)
 			require.Equal(t, state, InvalidState)
 		}
@@ -633,79 +631,4 @@ func TestE2E_ValidScenarios(t *testing.T) {
 		require.NoError(t, err)
 		fmt.Printf("%v TXs confirmed", instances)
 	})
-}
-func WaitForRequestState(expectedState string, ctx context.Context, chainID string, txHash string, requestURL string, apiKey string, timeout uint) (string, error) {
-	var (
-		previousState *BridgingRequestStateResponse
-		currentState  *BridgingRequestStateResponse
-		err           error
-	)
-
-	timeoutTimer := time.NewTimer(time.Second * time.Duration(timeout))
-	defer timeoutTimer.Stop()
-
-	for {
-		previousState = currentState
-
-		select {
-		case <-timeoutTimer.C:
-			fmt.Printf("Timeout\n")
-
-			return "", errors.New("Timeout")
-		case <-ctx.Done():
-			fmt.Printf("Done\n")
-
-			return "", errors.New("Done")
-		case <-time.After(time.Millisecond * 500):
-		}
-
-		currentState, err = GetBridgingRequestState(ctx, chainID, txHash, requestURL, apiKey, timeout)
-		if err != nil || currentState == nil {
-			continue
-		}
-
-		if previousState != nil && strings.Compare(previousState.Status, currentState.Status) != 0 {
-			fmt.Println(currentState)
-		}
-
-		if strings.Compare(currentState.Status, expectedState) == 0 {
-			return currentState.Status, nil
-		}
-	}
-}
-
-func GetBridgingRequestState(ctx context.Context, chainID string, txHash string, requestURL string, apiKey string, timeout uint) (*BridgingRequestStateResponse, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, requestURL, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("X-API-KEY", apiKey)
-	resp, err := http.DefaultClient.Do(req)
-
-	if resp == nil || err != nil || resp.StatusCode != http.StatusOK {
-		return nil, err
-	}
-
-	resBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	var responseModel *BridgingRequestStateResponse
-
-	err = json.Unmarshal(resBody, &responseModel)
-	if err != nil {
-		return nil, err
-	}
-
-	return responseModel, nil
-}
-
-type BridgingRequestStateResponse struct {
-	SourceChainID      string `json:"sourceChainId"`
-	SourceTxHash       string `json:"sourceTxHash"`
-	DestinationChainID string `json:"destinationChainId"`
-	Status             string `json:"status"`
-	DestinationTxHash  string `json:"destinationTxHash"`
 }
